Add :close-on-realign flag
If you open 5 different buffers that all match the "below" alignment,
they will stack and you'll see 5 windows.

Introduce a `:close-on-realign` flag, which will cause the window to be
deleted next time you display a window with the same alignment
direction. This has the effect that the new window replaces the previous
window, rather than stacking.